Moving Spandex into jembatan.core
